feat: Enhance model definitions and add new classes
- Updated `__init__.py` to include base model imports.
- Expanded `admin.py` with new classes: `Permissions`, `ACL`, `ParameterValue`, and `TrendingTag`.
- Modified existing classes to improve field definitions and default values.
- Introduced `StoryTag` class in `assess.py` for better tagging functionality.
- Added `ItemType` enum in `types.py` for categorizing items.
- Updated `base.py` to include a default model name for better clarity.
- Upgraded dependencies in `uv.lock` to latest versions for improved stability and features.
