+++ src/frontend/frontend/views/role_views.py
@@ -1,5 +1,5 @@
 # role_views.py
-from frontend.models import Role, Permissions
+from models.admin import Role, Permissions
 from frontend.data_persistence import DataPersistenceLayer
 from frontend.views.base_view import BaseView
