+++ src/frontend/frontend/views/worker_views.py
@@ -1,5 +1,5 @@
 # role_views.py
-from frontend.models import Worker
+from models.admin import Worker
 from frontend.views.base_view import BaseView
