+++ src/frontend/pyproject.toml
@@ -27,6 +27,7 @@ dependencies = [
     "requests",
     "setuptools",
     "heroicons[jinja]>=2.11.0",
+    "taranis-models",
 ]
 dynamic = ["version"]
 
@@ -76,3 +77,9 @@ exclude = [
     '__main__.py',
     'tests/'
 ]
+
+[tool.uv]
+config-settings = { editable_mode = "compat" }
+
+[tool.uv.sources]
+taranis-models = { path = "../models", editable = true }
