+++ src/gui/package.json
@@ -19,13 +19,13 @@
   "dependencies": {
     "@codemirror/lang-html": "^6.4.9",
     "@codemirror/lang-json": "^6.0.1",
-    "@intlify/unplugin-vue-i18n": "^6.0.5",
+    "@intlify/unplugin-vue-i18n": "^6.0.8",
     "@mdi/font": "^7.4.47",
-    "@sentry/vue": "^9.13.0",
+    "@sentry/vue": "^9.14.0",
     "@vue-js-cron/vuetify": "^5.0.9",
     "@vuepic/vue-datepicker": "^11.0.2",
     "@vueuse/core": "^13.1.0",
-    "axios": "^1.8.4",
+    "axios": "^1.9.0",
     "chart.js": "^4.4.9",
     "codemirror": "^6.0.1",
     "js-base64": "^3.7.7",
@@ -41,19 +41,19 @@
     "vue-codemirror6": "^1.3.13",
     "vue-dompurify-html": "^5.2.0",
     "vue-i18n": "^11.1.3",
-    "vue-router": "^4.5.0",
+    "vue-router": "^4.5.1",
     "vue-use-hotkeys": "^0.4.0",
-    "vuetify": "^3.8.2",
+    "vuetify": "^3.8.3",
     "zipson": "^0.2.12"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^5.2.3",
-    "eslint": "^9.24.0",
+    "eslint": "^9.25.1",
     "prettier": "^3.5.3",
-    "sass": "^1.86.3",
+    "sass": "^1.87.0",
     "typescript": "^5.8.3",
-    "vite": "^6.3.1",
+    "vite": "^6.3.3",
     "vite-plugin-vuetify": "^2.1.1",
-    "vue-tsc": "^2.2.8"
+    "vue-tsc": "^2.2.10"
   }
 }
