+++ src/models/models/__init__.py
@@ -1,5 +1,6 @@
 from models.admin import *  # noqa: F401
 from models.types import *  # noqa: F401
+from models.base import *  # noqa: F401
 
 __all__ = [
     "Job",
