+++ src/models/models/admin.py
@@ -1,5 +1,9 @@
+from pydantic import Field, AnyUrl
+from typing import Literal
+
 from models.base import TaranisBaseModel
-from models.types import TLPLevel
+from models.types import TLPLevel, ItemType
+from models.assess import StoryTag
 
 
 class Job(TaranisBaseModel):
@@ -11,56 +15,96 @@ class Job(TaranisBaseModel):
 
 
 class Address(TaranisBaseModel):
-    city: str | None = None
-    country: str | None = None
-    street: str | None = None
-    zip: str | None = None
+    city: str = ""
+    country: str = ""
+    street: str = ""
+    zip: str = ""
 
 
 class Organization(TaranisBaseModel):
     _core_endpoint = "/config/organizations"
+    _model_name = "organization"
     _search_fields = ["name", "description"]
 
     id: int | None = None
-    name: str
-    description: str | None = None
-    address: Address | None = None
-
-
-class Role(TaranisBaseModel):
-    _core_endpoint = "/config/roles"
-    _search_fields = ["name", "description"]
-
-    id: int | None = None
-    name: str
-    description: str | None = None
-    permissions: list[str] | None = None
-    tlp_level: TLPLevel | None = None
-
-
-class User(TaranisBaseModel):
-    _core_endpoint = "/config/users"
-    _search_fields = ["name", "username"]
-
-    id: int | None = None
-    name: str
-    organization: Organization | int | dict
-    permissions: list[str] | None = None
-    profile: dict | None = None
-    roles: list[Role] | list[int] | list[dict]
-    username: str
-    password: str | None = None
+    name: str = ""
+    description: str = ""
+    address: Address = Field(default_factory=Address)
 
 
 class Permissions(TaranisBaseModel):
     _core_endpoint = "/config/permissions"
+    _model_name = "permission"
     id: str
     name: str
     description: str
 
 
+class ACL(TaranisBaseModel):
+    _core_endpoint = "/config/acls"
+    _model_name = "acl"
+    _search_fields = ["name", "description"]
+
+    id: int | None = None
+    name: str = ""
+    description: str | None = ""
+    item_type: ItemType | None = None
+
+    roles: list["Role"] = Field(default_factory=list["Role"])
+
+    read_only: bool = True
+    enabled: bool = True
+
+
+class ParameterValue(TaranisBaseModel):
+    _core_endpoint = "/config/parameter-values"
+    _model_name = "parameter_value"
+    id: int | None = None
+    parameter: str = ""
+    value: str | None = ""
+
+
+class Worker(TaranisBaseModel):
+    _core_endpoint = "/config/workers"
+    _model_name = "worker"
+    id: str
+    name: str
+    description: str | None = ""
+    type: str | None = ""
+    category: str | None = ""
+    parameters: list["ParameterValue"] = Field(default_factory=list["ParameterValue"])
+
+
+class Role(TaranisBaseModel):
+    _core_endpoint = "/config/roles"
+    _model_name = "role"
+    _search_fields = ["name", "description"]
+
+    id: int | None = None
+    name: str = ""
+    description: str | None = ""
+    permissions: list[str] = Field(default_factory=list[str])
+    tlp_level: TLPLevel | None = None
+
+
+class User(TaranisBaseModel):
+    _core_endpoint = "/config/users"
+    _model_name = "user"
+    _search_fields = ["name", "username"]
+
+    id: int | None = None
+    name: str = ""
+    organization: Organization | int | dict = Field(default_factory=dict)
+    permissions: list[str] | None = None
+    profile: dict | None = None
+    roles: list[Role] | list[int] | list[dict] = Field(default_factory=list[Role])
+    username: str = ""
+    password: str | None = None
+
+
 class Dashboard(TaranisBaseModel):
     _core_endpoint = "/dashboard"
+    _model_name = "dashboard"
     _cache_timeout = 30
     total_news_items: int | None = None
     total_products: int | None = None
@@ -69,3 +113,32 @@ class Dashboard(TaranisBaseModel):
     total_database_items: int | None = None
     latest_collected: str | None = None
     schedule_length: int | None = None
+    conflict_count: int | None = None
+
+
+class TrendingTag(TaranisBaseModel):
+    _core_endpoint = "/dashboard/cluster"
+    _model_name = "trending_tags"
+    name: str
+    tags: list[StoryTag] = Field(default_factory=list)
+    size: int | None = None
+
+
+class TrendingClusters(TaranisBaseModel):
+    _core_endpoint = "/dashboard/trending-clusters"
+    _model_name = "trending_clusters"
+    root: list[TrendingTag] = Field(default_factory=list)
+
+
+class TaranisConfig(TaranisBaseModel):
+    default_collector_proxy: AnyUrl | None | Literal[""] = None
+    default_collector_interval: str | None = None
+    default_tlp_level: TLPLevel | None = None
+
+
+class Settings(TaranisBaseModel):
+    _core_endpoint = "/admin/settings"
+    _model_name = "settings"
+    _cache_timeout = 30
+    id: int = Field(default=1, frozen=True, exclude=True)
+    settings: TaranisConfig | None = None
