+++ src/models/models/base.py
@@ -7,6 +7,7 @@ T = TypeVar("T", bound="TaranisBaseModel")
 class TaranisBaseModel(BaseModel):
     _core_endpoint: ClassVar[str]
     _cache_timeout: ClassVar[int]
+    _model_name: ClassVar[str] = ""
 
     def model_dump(self, *args, **kwargs):
         kwargs.setdefault("exclude_none", True)
