+++ src/models/models/types.py
@@ -1,4 +1,4 @@
-from enum import StrEnum
+from enum import StrEnum, auto
 
 
 class TLPLevel(StrEnum):
@@ -7,3 +7,11 @@ class TLPLevel(StrEnum):
     AMBER_STRICT = "amber+strict"
     AMBER = "amber"
     RED = "red"
+
+
+class ItemType(StrEnum):
+    OSINT_SOURCE = auto()
+    OSINT_SOURCE_GROUP = auto()
+    WORD_LIST = auto()
+    REPORT_ITEM_TYPE = auto()
+    PRODUCT_TYPE = auto()
